2025/aiogram-webhooks-1/pyproject.toml
[project]
name = "aiogram-webhooks-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.11.0",
    "aiohttp>=3.9.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "importlib-metadata>=6.8.0; python_version < \"3.10\"",
]
